fix: cppcheck narrowingConversion — size_t to int in latency benchmarks
Seven demo files assigned latencies.size() (size_t, 64-bit) into int n
(32-bit signed) without an explicit cast. cppcheck --enable=warning flags
this as implicit narrowing. Fix: static_cast<int>(latencies.size()).

Files fixed: risk_demo, router_demo, strategy_demo, fix_demo, ouch_demo,
logger_demo, oms_demo.

+++ ouch-protocol/ouch_demo.cpp
@@ -194,7 +194,7 @@ void benchmark(int num_encodes) {
         total_end - total_start).count();
 
     std::sort(latencies.begin(), latencies.end());
-    int n = latencies.size();
+    int n = static_cast<int>(latencies.size());
     double avg = static_cast<double>(total_ns) / n;
     double throughput = n / (total_ns / 1e9) / 1e6;
