+++ router/router_demo.cpp
@@ -140,7 +140,7 @@ void benchmark(int num_routes) {
         total_end - total_start).count();
 
     std::sort(latencies.begin(), latencies.end());
-    int n = latencies.size();
+    int n = static_cast<int>(latencies.size());
     double avg = static_cast<double>(total_ns) / n;
     double throughput = n / (total_ns / 1e9) / 1e6;
